refactor(connect): keep authenticateDevice runtime schema out of connect-common
@trezor/connect-common previously imported DeviceAuthenticityConfig and
DeviceAuthenticityBlacklistConfig as runtime values from
@trezor/device-authenticity to construct AuthenticateDeviceParams as a
TypeBox schema. That made @trezor/device-authenticity a runtime
dependency of connect-common — and therefore a transitive runtime
dependency of every thin public connect package
(@trezor/connect-web, @trezor/connect-webextension, @trezor/connect-mobile),
which never need the device-authenticity verification logic.

This change makes @trezor/device-authenticity the single source of truth
for the AuthenticateDeviceParams schema. The TypeBox schema and its
Static<typeof ...> type are now declared together in
device-authenticity/src/authenticateDeviceParams.ts, sharing one name
the way DeviceAuthenticityConfig and DeviceAuthenticityBlacklistConfig
already do. The runtime closure lines up with the domain: every field
of the schema (config, blacklistConfig, allowDebugKeys) is a
device-authenticity concept.

@trezor/connect-common's authenticateDevice type module becomes a
type-only re-export. The published bundle has zero runtime references
to @trezor/device-authenticity (the type is inlined as a vendor file by
inline-devdep-types.mjs). @trezor/connect imports the schema directly
and uses it both as the type and as the runtime Assert target — one
import, no duplicated shape, no drift.

@trezor/device-authenticity is moved back from connect-common's
dependencies to devDependencies (reverting the tactical fix in
63f9eb899d5 / #27401), and its tsconfig project references are
relocated to the devDeps block. The connect-{web,webextension,mobile}
public-package-dependencies snapshots drop @noble/curves,
@noble/post-quantum, @trezor/crypto-utils and @trezor/device-authenticity
— the four entries that #27401 added to the thin packages' runtime
closure.

@trezor/connect-explorer's authenticateDevice.mdx is updated to import
AuthenticateDeviceParams from @trezor/device-authenticity (it uses the
schema as a runtime value to render the ParamsTable), and the
device-authenticity workspace dep + tsconfig project reference are
added.

refs #27404
refs #27401
