+++ packages/connect-common/src/types/api/authenticateDevice.ts
@@ -1,19 +1,11 @@
-import type { VerifyAuthenticityProofResult } from '@trezor/device-authenticity';
-import {
-    DeviceAuthenticityBlacklistConfig,
-    DeviceAuthenticityConfig,
+import type {
+    AuthenticateDeviceParams,
+    VerifyAuthenticityProofResult,
 } from '@trezor/device-authenticity';
-import type { Static } from '@trezor/schema-utils';
-import { Type } from '@trezor/schema-utils';
 
 import type { Params, Response } from '../params';
 
-export type AuthenticateDeviceParams = Static<typeof AuthenticateDeviceParams>;
-export const AuthenticateDeviceParams = Type.Object({
-    config: Type.Optional(DeviceAuthenticityConfig),
-    blacklistConfig: Type.Optional(DeviceAuthenticityBlacklistConfig),
-    allowDebugKeys: Type.Optional(Type.Boolean()),
-});
+export type { AuthenticateDeviceParams };
 
 export type AuthenticateDeviceResult = {
     // Signed by Optiga secure element, available in all Trezor Safe devices.
