+++ packages/connect-common/tsconfig.json
@@ -6,7 +6,6 @@
     },
     "include": ["."],
     "references": [
-        { "path": "../device-authenticity" },
         { "path": "../device-utils" },
         { "path": "../protobuf" },
         { "path": "../protocol" },
@@ -15,6 +14,7 @@
         { "path": "../utils" },
         { "path": "../blockchain-link" },
         { "path": "../blockchain-link-types" },
+        { "path": "../device-authenticity" },
         { "path": "../transport" },
         { "path": "../utxo-lib" }
     ]
