+++ packages/connect-common/tsconfig.libESM.json
@@ -7,7 +7,6 @@
         "types": ["node", "chrome", "jest"]
     },
     "references": [
-        { "path": "../device-authenticity" },
         { "path": "../device-utils" },
         { "path": "../protobuf" },
         { "path": "../protocol" },
@@ -15,6 +14,7 @@
         { "path": "../utils" },
         { "path": "../blockchain-link" },
         { "path": "../blockchain-link-types" },
+        { "path": "../device-authenticity" },
         { "path": "../transport" },
         { "path": "../utxo-lib" }
     ]
