+++ packages/connect-explorer/src/pages/methods/device/authenticateDevice.mdx
@@ -1,6 +1,6 @@
 import { Callout } from 'nextra/components';
 
-import { AuthenticateDeviceParams } from '@trezor/connect-common/src/types/api/authenticateDevice';
+import { AuthenticateDeviceParams } from '@trezor/device-authenticity';
 
 import { ApiPlayground } from '../../../components/ApiPlayground';
 import { CommonParamsLink } from '../../../components/CommonParamsLink';
